refactor: Move to session/ui/detail dir
