+++ lib/features/session/ui/detail/session_detail.dart
@@ -3,7 +3,7 @@ import 'package:confwebsite2023/core/components/responsive_widget.dart';
 import 'package:confwebsite2023/core/theme.dart';
 import 'package:confwebsite2023/features/session/data/session_model.dart';
 import 'package:confwebsite2023/features/session/data/talk_user.dart';
-import 'package:confwebsite2023/features/session/ui/session_detail_content.dart';
+import 'package:confwebsite2023/features/session/ui/detail/session_detail_content.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
