+++ lib/features/session/ui/detail/session_detail_content.dart
